Map reading OK

+++ src/download_bizhawk.sh
@@ -2,14 +2,26 @@
 
 set -e
 
+isWsl=$(grep -i microsoft /proc/version)
+
+if [[ "$isWsl" ]]; then
+  echo "WSL detected"
+  file=BizHawk-2.6.2-win-x64.zip
+else
+  echo "Native linux detected"
+  file=BizHawk-2.6.2-linux-x64.tar.zip
+fi
+
 pushd $(dirname $0)/..
-if [[ -z Bizhawk.tar.zip ]]; then
-  wget https://github.com/TASEmulators/BizHawk/releases/download/2.6.2/BizHawk-2.6.2-linux-x64.tar.zip -O BizHawk.tar.zip
+if [[ ! -f $file ]]; then
+  wget "https://github.com/TASEmulators/BizHawk/releases/download/2.6.2/$file"
 fi
 rm -rf BizHawk
-mkdir BizHawk 
-unzip BizHawk.tar.zip -d BizHawk
+mkdir BizHawk
+unzip "$file" -d BizHawk
 cd BizHawk
-tar xf BizHawk_release_*.tar
-rm BizHawk_release_*.tar
+if [[ ! "$isWsl" ]]; then
+  tar xf BizHawk_release_*.tar
+  rm BizHawk_release_*.tar
+fi
 popd
